BUG: Fix `TensorSlicerActor` actor opacity property not being enforced.
Fix `TensorSlicerActor` actor opacity property not being enforced.

The method `tensor_slicer` returns an instance of a `TensorSlicerActor`
actor, and receives, among others the `opacity` that needs to be applied
to the actor as a parameter. However, the parameter was being passed to
the `_tensor_slicer_mapper` method, which returns a mapper, and thus was
not effectively using the `opacity` property.

Now the `opacity` parameter is no longer passed to the
`_tensor_slicer_mapper` and it is set as a `vtk::Property` to the
`TensorSlicerActor` instance.

The corresponding actor test is modified so that the `opacity` property
is effectively exercised.