improves code organization and readability using abstract classes
After this commit it will be easier to define new memory management objs.
For example, through socket communication (py3.7 compatibility).

In addition, this commit makes the code more closely to SOLID.
For example, before this commit the FuryStreamClient was responsible
to create the shared memory, resources, start the vtk callbacks and
more.