STYLE: Transition to `ruff` to enforce import statement sorting
Prefer using `ruff` instead of `isort` to enforce import statement
sorting.

`ruff`'s import sorting rules are near-equivalent to isort's when using
`isort`'s `profile="black"`:
https://docs.astral.sh/ruff/faq/#how-does-ruffs-import-sorting-compare-to-isort

Change the `tool.isort` section in `pyproject` so that the correspoding
tasks fall under the responsibility of `flake8`.

Adapt accordingly the pre-commit hook config file to apply the import
sorting to local files prior to commits.

Add a GHA workflow to apply the pre-commit hooks in order to ensure that
files modified across commits/in PRs comply with the expected
formatting.

Add the `--allow-missing-credentials` flag to the
`detect-aws-credentials` hook in the `pre-commit` config file so that
`pre-commit` does not complain about not finding AWS credentials if none
exist.

Make the necessary style changes to make the `ruff` linter happy.